chore: replace abstract class with an interface
