+++ packages/idos-sdk-js/src/lib/enclave-providers/interface.ts
@@ -1,0 +1,21 @@
+export interface StoredData {
+  encryptionPublicKey?: string;
+  humanId?: string;
+  signerAddress?: string;
+  signerPublicKey?: string;
+}
+
+export interface EnclaveProvider {
+  load(): Promise<StoredData>;
+  init(
+    humanId?: string,
+    signerAddress?: string,
+    signerPublicKey?: string,
+    authMethod?: boolean
+  ): Promise<Uint8Array>;
+  store(key: string, value: string): Promise<string>;
+  reset(): Promise<void>;
+  confirm(message: string): Promise<boolean>;
+  encrypt(message: Uint8Array, receiverPublicKey?: Uint8Array): Promise<Uint8Array>;
+  decrypt(message: Uint8Array, senderPublicKey?: Uint8Array): Promise<Uint8Array>;
+}
