+++ packages/idos-sdk-js/src/lib/index.ts
@@ -1,5 +1,5 @@
 export * from "./idos";
 export { default as Grant } from "./grants/grant";
-export * from "./enclave-providers/enclave-provider";
+export * from "./enclave-providers/interface";
 export * from "./enclave-providers";
 export * from "./utils";
